Get Vico compiling on the 10.8 SDK and update Nu to 2.1.

+++ app/ViTaskRunner.m
@@ -151,7 +151,9 @@
 	const void *ptr;
 	NSUInteger len;
 
-	switch (event) {
+	// cast to int because ViStreamEventWriteEndEncountered is not declared
+	// part of the enum, and clang don't like that
+	switch ((int)event) {
 	case NSStreamEventNone:
 	case NSStreamEventOpenCompleted:
 	default:
